+++ pdr_backend/predictoor/approach3/predictoor_agent3.py
@@ -1,3 +1,4 @@
+import copy
 from typing import Tuple
 
 from enforce_typing import enforce_types
